feat(api): initial api integration architecture design
