+++ packages/api/start/jobs.js
@@ -2,5 +2,6 @@ module.exports = [
 	'App/Jobs/TechnologyDistribution',
 	'App/Jobs/SendMail',
 	'App/Jobs/UpdateTechnologyTotalViews',
-	'App/Jobs/UpdateElasticMetrics',
+	'App/Jobs/ElasticSearch/Publisher',
+	'App/Jobs/ElasticSearch/Consumer',
 ];
